+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "gridmanager-angular-1.x",
-    "version": "0.0.4",
+    "version": "0.0.5",
     "description": "gridmanager by angular-1.x",
     "main": "js/gm-angular.js",
     "scripts": {
@@ -30,15 +30,12 @@
     ],
     "devDependencies": {
         "angular": "^1.7.5",
-        "angular-resource": "^1.7.5",
-        "angular-ui-router": "^1.0.20",
         "babel-cli": "6.24.1",
         "babel-core": "6.25.0",
         "babel-loader": "6.4.1",
         "babel-plugin-transform-decorators-legacy": "1.3.4",
         "babel-preset-es2015": "6.24.1",
         "babel-preset-stage-0": "6.24.1",
-        "ccms-components": "^2.35.12",
         "copy-webpack-plugin": "^4.5.0",
         "css-loader": "^0.28.10",
         "express": "^4.16.3",
