+++ src/demo/index.html
@@ -4,28 +4,13 @@
     <meta charset="UTF-8">
     <title>gm-angular</title>
     <link type="text/css" rel="stylesheet" href="style.css"/>
-    <link rel="stylesheet" href="../../node_modules/ccms-components/ccms-components.css">
     <script src="../../node_modules/angular/angular.min.js"></script>
-    <script src="../../node_modules/ccms-components/ccms-components.js"></script>
 
     <link type="text/css" rel="stylesheet" href="/webpack-dev-file/css/gm-angular.css"/>
     <script src="/webpack-dev-file/js/gm-angular.js"></script>
     <style>
         .grid-main {
             height: calc(100vh - 64px - 57px);
-        }
-        table[grid-manager="testAngular"] th{
-            padding: 0;
-        }
-        table[grid-manager="testAngular"] th .th-wrap{
-            padding: 0 !important;
-        }
-
-        table[grid-manager="testAngular"] .ffff{
-            display: flex;
-        }
-        table[grid-manager="testAngular"] .ffff p{
-            flex-basis: 33%;
         }
     </style>
 </head>
